extract formatMoney into helper
